pyproject.toml
[tool.poetry]
name = "fcc-fastapi"
version = "0.1.0"
description = ""
authors = ["Lambov"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.112.2"}
sqlalchemy = "^2.0.32"
psycopg2-binary = "^2.9.9"
passlib = "^1.7.4"
pyjwt = "^2.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
